Changed swizzle indices to size_t (from int).

+++ src/math/core/swizzle.hpp
@@ -1,6 +1,8 @@
 #ifndef MATH_CORE_SWIZZLE_HPP
 #define MATH_CORE_SWIZZLE_HPP
 
+#include <cstddef>
+
 namespace math {
 namespace core {
 
@@ -18,7 +20,7 @@ class vector4;
  * @tparam T Storage type.
  * @tparam N Storage size.
  */
-template<typename T, int N>
+template<typename T, size_t N>
 class Swizzle {
 protected:
 	T values[N];
@@ -31,7 +33,7 @@ protected:
  * @tparam E1 Index of first element.
  * @tparam E2 Index of second element.
  */
-template<typename T, int N, int E1, int E2>
+template<typename T, size_t N, size_t E1, size_t E2>
 class Swizzle2 : Swizzle<T, N> {
 public:
 	operator vector2<T>();
@@ -50,7 +52,7 @@ public:
  * @tparam E2 Index of second element.
  * @tparam E3 Index of third element.
  */
-template<typename T, int N, int E1, int E2, int E3>
+template<typename T, size_t N, size_t E1, size_t E2, size_t E3>
 class Swizzle3 : Swizzle<T, N> {
 public:
 	operator vector3<T>();
@@ -70,7 +72,7 @@ public:
  * @tparam E3 Index of third element.
  * @tparam E4 Index of fourth element.
  */
-template<typename T, int N, int E1, int E2, int E3, int E4>
+template<typename T, size_t N, size_t E1, size_t E2, size_t E3, size_t E4>
 class Swizzle4 : Swizzle<T, N> {
 public:
 	operator vector4<T>();
